src/agents_resolvers/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.38.29